Points can be more than integers can hold, so they are now longs

+++ app/src/main/java/com/imgurisnotimgur/entities/Image.kt
@@ -8,7 +8,7 @@ class Image (
         val id: String,
         val title: String,
         val username: String,
-        val points: Int,
+        val points: Long,
         val createdAt: Long,
         val albumId: String,
         val isAlbum: Boolean
